refactor: make FlipCard and ShareableProfileCard more stateless
